+++ CI/SCRIPTS/test_unistra.py
@@ -34,7 +34,7 @@ from sertit.unistra import (
 def test_unistra_s3():
     with tempenv.TemporaryEnvironment({SU_USE_S3: "1", CI_SERTIT_S3: "1"}):
         # Test s3_env and define_s3_client (called inside)
-        def base_fct():
+        def base_fct(value):
             raster_path = AnyPath("s3://sertit-sertit-utils-ci").joinpath(
                 "DATA", "rasters", "raster.tif"
             )
@@ -45,11 +45,11 @@ def test_unistra_s3():
 
         @s3_env
         def with_s3():
-            base_fct()
+            base_fct(1)
 
         def without_s3():
             S3Client().set_as_default_client()
-            base_fct()
+            base_fct(None)
 
         with pytest.raises(AssertionError):
             without_s3()